+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/management/infrastructure/web/controller/StatusListIT.java
@@ -11,7 +11,6 @@ import ch.admin.bj.swiyu.issuer.common.config.StatusListProperties;
 import ch.admin.bj.swiyu.issuer.common.config.SwiyuProperties;
 import ch.admin.bj.swiyu.issuer.domain.credentialoffer.StatusList;
 import ch.admin.bj.swiyu.issuer.domain.credentialoffer.StatusListRepository;
-import ch.admin.bj.swiyu.issuer.domain.credentialoffer.StatusListType;
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerMetadata;
 import ch.admin.bj.swiyu.issuer.service.JwsSignatureFacade;
 import ch.admin.bj.swiyu.jwssignatureservice.factory.strategy.KeyStrategyException;
@@ -141,14 +140,15 @@ class StatusListIT {
 
     @Test
     void createNewStatusListOverrideConfiguration_thenSuccess() throws Exception {
-        final StatusListType type = StatusListType.TOKEN_STATUS_LIST;
+
         final int maxLength = 127;
         final int bits = 4;
         final String issuerId = "did:example:offer:override";
         final String verificationMethod = issuerId + "#key";
         final String keyId = "1052933";
         final String keyPin = "209323";
-        final String payload = String.format("{\"type\": \"%s\",\"maxLength\": %d,\"config\": {\"bits\": %d},\"configuration_override\": {\"issuer_did\": \"%s\",\"verification_method\": \"%s\",\"key_id\": %s,\"key_pin\": %s}}", type, maxLength, bits, issuerId, verificationMethod, keyId, keyPin);
+        final String payload = String.format(
+                "{\"type\": \"%s\",\"maxLength\": %d,\"config\": {\"bits\": %d},\"configuration_override\": {\"issuer_did\": \"%s\",\"verification_method\": \"%s\",\"key_id\": %s,\"key_pin\": %s}}", maxLength, bits, issuerId, verificationMethod, keyId, keyPin);
 
         MvcResult result = mvc.perform(post(STATUS_LIST_BASE_URL)
                         .contentType(MediaType.APPLICATION_JSON)
@@ -162,7 +162,6 @@ class StatusListIT {
         assertTrue(newStatusListOpt.isPresent());
         final StatusList newStatusList = newStatusListOpt.get();
         assertNotNull(newStatusList.getUri());
-        assertEquals(type, newStatusList.getType());
         assertEquals(maxLength, newStatusList.getMaxLength());
         assertEquals(bits, newStatusList.getConfig().get("bits"));
         assertEquals(issuerId, newStatusList.getConfigurationOverride().issuerDid());