+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/common/config/ApplicationProperties.java
@@ -51,9 +51,6 @@ public class ApplicationProperties {
     @NotNull
     private List<String> trustedAttestationProviders;
 
-    @NotNull
-    private String requestOfferVersion = "1.0";
-
     /**
      * If set to true the service expects all
      * writing message bodies to be encoded as JWT