+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/dto/statuslist/StatusListTypeDto.java
@@ -1,8 +1,0 @@
-package ch.admin.bj.swiyu.issuer.dto.statuslist;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(name = "StatusListType")
-public enum StatusListTypeDto {
-    TOKEN_STATUS_LIST
-}