+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/offer/CredentialOfferMapper.java
@@ -175,7 +175,6 @@ public class CredentialOfferMapper {
                 .credentialIssuer(credentialIssuer)
                 .credentials(credentialOffer.getMetadataCredentialSupportedId())
                 .grants(grants)
-                .version(props.getRequestOfferVersion())
                 .build();
 
         String credentialOfferString;