+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/statuslist/StatusListOrchestrator.java
@@ -4,7 +4,6 @@ import ch.admin.bj.swiyu.core.status.registry.client.model.StatusListEntryCreati
 import ch.admin.bj.swiyu.issuer.dto.credentialoffer.CreateCredentialOfferRequestDto;
 import ch.admin.bj.swiyu.issuer.dto.statuslist.StatusListCreateDto;
 import ch.admin.bj.swiyu.issuer.dto.statuslist.StatusListDto;
-import ch.admin.bj.swiyu.issuer.dto.statuslist.StatusListTypeDto;
 import ch.admin.bj.swiyu.issuer.common.config.ApplicationProperties;
 import ch.admin.bj.swiyu.issuer.common.config.StatusListProperties;
 import ch.admin.bj.swiyu.issuer.common.exception.BadRequestException;
@@ -89,8 +88,7 @@ public class StatusListOrchestrator {
                 .orElseThrow(() -> new ResourceNotFoundException(String.format("Status List %s not found", statusListId)));
 
         return toStatusListDto(statusList,
-                statusList.getMaxLength() - credentialOfferStatusRepository.countByStatusListId(statusListId),
-                statusListProperties.getVersion());
+                statusList.getMaxLength() - credentialOfferStatusRepository.countByStatusListId(statusListId));
     }
 
     /**
@@ -106,14 +104,13 @@ public class StatusListOrchestrator {
             // use explicit transaction, since we want to handle data integrity exceptions
             // after commit
             var newStatusList = transaction.execute(status -> {
-                var statusListType = request.getType();
-                var statusList = switch (statusListType) {
-                    case TOKEN_STATUS_LIST -> initTokenStatusListToken(request);
-                };
+
+                var statusList = initTokenStatusListToken(request);
+
                 return statusListRepository.save(statusList);
             });
 
-            return toStatusListDto(newStatusList, newStatusList.getMaxLength(), statusListProperties.getVersion());
+            return toStatusListDto(newStatusList, newStatusList.getMaxLength());
         } catch (DataIntegrityViolationException e) {
             var msg = e.getMessage();
             if (msg != null && msg.toLowerCase().contains("status_list_uri_key")) {
@@ -159,8 +156,7 @@ public class StatusListOrchestrator {
         publishToRegistry(statusList, token);
 
         return toStatusListDto(statusList,
-                statusList.getMaxLength() - credentialOfferStatusRepository.countByStatusListId(statusList.getId()),
-                statusListProperties.getVersion());
+                statusList.getMaxLength() - credentialOfferStatusRepository.countByStatusListId(statusList.getId()));
     }
 
 
@@ -201,7 +197,6 @@ public class StatusListOrchestrator {
 
         // Build DB Entry
         StatusList statusList = StatusList.builder()
-                .type(getStatusListTypeFromDto(statusListCreateDto.getType()))
                 .config(Map.of(
                         BITS_FIELD_NAME, config.getBits(),
                         "purpose", config.getPurpose() != null ? config.getPurpose() : ""
@@ -214,14 +209,6 @@ public class StatusListOrchestrator {
         log.debug("Initializing new status list with bit {} per entry and {} entries to a total size of {} bit", config.getBits(), statusList.getMaxLength(), config.getBits() * statusList.getMaxLength());
         publishToRegistry(statusList, token);
         return statusList;
-    }
-
-    private StatusListType getStatusListTypeFromDto(StatusListTypeDto statusListTypeDto) {
-        if (statusListTypeDto == null) {
-            return null;
-        }
-
-        return StatusListType.TOKEN_STATUS_LIST;
     }
 
     private StatusListEntryCreationDto createEmptyRegistryEntry() {